+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Updated holidays
 - Changed warning to error for unsupported country holidays
 - Disabled logging by default
+- Fixed error with `add_regressor` and holidays
 - Dropped support for Ruby < 3
 
 ## 0.4.2 (2022-07-12)