+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.5.0 (unreleased)
+## 0.5.0 (2023-09-05)
 
 - Added support for Polars
 - Updated holidays